+++ CHANGES.rst
@@ -3,6 +3,10 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
+**UNRELEASED**
+
+- Patch ``sys.unraisablehook`` to properly format exceptiongroups
+
 **1.0.0rc8**
 
 - Don't monkey patch anything if ``sys.excepthook`` has been altered